refactor(component): split vdom and hydration test suites
Break up the large vdom and hydration spec tests into smaller, topic-focused files. There's a ton of behavior to assert with a small amount of public API, moved everything into test/ because it was hard to find the module you want in lib/ with all the test files next to them
