+++ packages/component/src/test/create-element.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest'
 
-import { createElement } from './create-element.ts'
+import { createElement } from '../lib/create-element.ts'
 
 describe('createElement', () => {
   it('creates an element', () => {
