+++ packages/component/src/test/vdom.svg.test.tsx
@@ -1,0 +1,157 @@
+import { describe, it, expect } from 'vitest'
+import { createRoot } from '../lib/vdom.ts'
+import { invariant } from '../lib/invariant.ts'
+import type { Handle, RemixNode } from '../lib/component.ts'
+
+describe('vnode rendering', () => {
+  describe('svg', () => {
+    it('renders SVG root and children with SVG namespace and attributes', () => {
+      let container = document.createElement('div')
+      let { render } = createRoot(container)
+
+      render(
+        <svg viewBox="0 0 24 24" fill="none">
+          <path strokeLinecap="round" strokeLinejoin="round" d="m4.5 12.75 6 6 9-13.5" />
+        </svg>,
+      )
+
+      let svg = container.querySelector('svg')
+      let path = container.querySelector('path')
+      invariant(svg instanceof SVGSVGElement)
+      invariant(path instanceof SVGPathElement)
+
+      expect(svg.namespaceURI).toBe('http://www.w3.org/2000/svg')
+      expect(path.namespaceURI).toBe('http://www.w3.org/2000/svg')
+
+      // Attribute casing: preserve exceptions and kebab-case general SVG attrs
+      expect(svg.getAttribute('viewBox')).toBe('0 0 24 24')
+      expect(path.getAttribute('stroke-linecap')).toBe('round')
+      expect(path.getAttribute('stroke-linejoin')).toBe('round')
+    })
+
+    it('supports xlinkHref -> xlink:href on SVG elements', () => {
+      let container = document.createElement('div')
+      let { render } = createRoot(container)
+
+      render(
+        <svg>
+          <use xlinkHref="#my-id" />
+        </svg>,
+      )
+
+      let useEl = container.querySelector('use')
+      invariant(useEl instanceof SVGUseElement)
+
+      expect(useEl.getAttribute('xlink:href')).toBe('#my-id')
+    })
+
+    it('renders HTML subtree inside foreignObject with HTML namespace', () => {
+      let container = document.createElement('div')
+      let { render } = createRoot(container)
+
+      render(
+        <svg>
+          <foreignObject>
+            <div id="x">Hello</div>
+          </foreignObject>
+        </svg>,
+      )
+
+      let div = container.querySelector('#x')
+      invariant(div)
+      expect(div instanceof HTMLDivElement).toBe(true)
+      expect(div.namespaceURI).toBe('http://www.w3.org/1999/xhtml')
+    })
+
+    it('updates and removes SVG attributes', () => {
+      let container = document.createElement('div')
+      let root = createRoot(container)
+
+      root.render(
+        <svg>
+          <path id="p" strokeLinecap="round" />
+        </svg>,
+      )
+      let path = container.querySelector('#p')
+      invariant(path instanceof SVGPathElement)
+
+      // Update value
+      root.render(
+        <svg>
+          <path id="p" strokeLinecap="square" />
+        </svg>,
+      )
+      let updated = container.querySelector('#p')
+      invariant(updated instanceof SVGPathElement)
+      expect(updated).toBe(path)
+      expect(updated.getAttribute('stroke-linecap')).toBe('square')
+
+      // Remove attribute
+      root.render(
+        <svg>
+          <path id="p" />
+        </svg>,
+      )
+      let removed = container.querySelector('#p')
+      invariant(removed instanceof SVGPathElement)
+      expect(removed.hasAttribute('stroke-linecap')).toBe(false)
+    })
+
+    it('attaches events on SVG elements', () => {
+      let container = document.createElement('div')
+      let root = createRoot(container)
+
+      let clicked = false
+      root.render(
+        <svg>
+          <circle
+            id="c"
+            on={{
+              click: () => {
+                clicked = true
+              },
+            }}
+          />
+        </svg>,
+      )
+      root.flush() // attach events
+
+      let circle = container.querySelector('#c')
+      invariant(circle instanceof SVGCircleElement)
+      circle.dispatchEvent(new MouseEvent('click', { bubbles: true }))
+      expect(clicked).toBe(true)
+    })
+
+    it('propagates SVG namespace through components', () => {
+      let container = document.createElement('div')
+      let root = createRoot(container)
+
+      function SvgGroup() {
+        return ({ href, children }: { href: string; children?: RemixNode }) => (
+          <g href={href}>{children}</g>
+        )
+      }
+
+      root.render(
+        <svg width="100" height="100">
+          <SvgGroup href="/test">
+            <path id="p" />
+          </SvgGroup>
+        </svg>,
+      )
+
+      let svg = container.querySelector('svg')
+      let group = container.querySelector('g')
+      let path = container.querySelector('path')
+
+      invariant(svg instanceof SVGSVGElement)
+      invariant(group instanceof SVGGElement)
+      invariant(path instanceof SVGPathElement)
+
+      // All elements should have SVG namespace
+      expect(svg.namespaceURI).toBe('http://www.w3.org/2000/svg')
+      expect(group.namespaceURI).toBe('http://www.w3.org/2000/svg')
+      expect(path.namespaceURI).toBe('http://www.w3.org/2000/svg')
+    })
+  })
+})
